IDETECT-793: clang refactoring: simplify extractor
